Create separate files for route utils
