+++ src/domain/app/routes/AppRoutes.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { Redirect, Route, Switch } from 'react-router';
 
-import { ResetFocus } from '../../../common/route/RouteUtils';
 import { getCurrentLanguage } from '../../../common/translation/TranslationUtils';
 import { SUPPORT_LANGUAGES } from '../../../constants';
 import { MobileMenuProvider } from '../header/mobileMenu/MobileMenu';
 import PageLayout from '../layout/PageLayout';
+import ResetFocus from '../resetFocus/ResetFocus';
 import App from './LocaleRoutes';
 
 const localeParam = `:locale(${SUPPORT_LANGUAGES.EN}|${SUPPORT_LANGUAGES.FI}|${SUPPORT_LANGUAGES.SV})`;
