+++ src/util/updateLocaleParam.ts
@@ -1,0 +1,7 @@
+export const updateLocaleParam = (
+  url: string,
+  currentLocale: string,
+  value: string
+): string => {
+  return url.replace(`/${currentLocale}/`, `/${value}/`);
+};
